+++ crates/ruma-client-api/CHANGELOG.md
@@ -44,6 +44,7 @@ Improvements:
 - Add server support discovery endpoint, according to MSC1929 / Matrix 1.10
 - Add `dir` `Request` field on the `get_relating_events_with_rel_types` and
   `get_relating_events_with_rel_type_and_event_type` endpoints
+- Add unstable support for moderator server support discovery, according to MSC4121
 
 # 0.17.4
 
@@ -83,7 +84,7 @@ Breaking changes:
 
 Improvements:
 
-- Add convenience constructors for enabling lazy-loading in filters 
+- Add convenience constructors for enabling lazy-loading in filters
 - Add support for using an existing session to log in another (MSC3882 / Matrix 1.7)
 - Add support for media download redirects (MSC3860 / Matrix 1.7)
 - Stabilize support for asynchronous media uploads (MSC2246 / Matrix 1.7)